+++ docs/changelog.rst
@@ -2,8 +2,10 @@
 Changelog
 =========
 
+
 * :feature:`436` Rotki now supports Bitstamp. Users can see their balances and import trades, deposits and withdrawals from that exchange. They are also taken into account in the tax report.
 * :feature:`1611` Rotki can now import data and download the tax report csv when running in the browser.
+* :feature:`1413` Users can now refresh their manual balances from the dashboard.
 * :feature:`176` Add an accounting setting to make asset movements fees (deposits/withdrawals to/from exchanges) inclusion in the profit loss report configurable.
 * :feature:`1840` Better handling double crypto.com entries (dust_conversion, swap, ...) from csv export. Also crypto.com imported trades and asset movements now appear in the history UI component
 * :feature:`1605` User funds in Binance's futures wallet should now also be included in Rotki